+++ Common/src/main/java/io/github/sirjain0/perfectplushies/block/DumboBlobPlushieBlock.java
@@ -5,9 +5,7 @@ import io.github.sirjain0.perfectplushies.block.entity.PPDualColorBlockEntity;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
-import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.EntityBlock;
 import net.minecraft.world.level.block.RenderShape;
@@ -15,7 +13,6 @@ import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.BlockHitResult;
 import org.jetbrains.annotations.Nullable;
-import software.bernie.geckolib.core.animation.AnimationController;
 
 public class DumboBlobPlushieBlock extends RandomDualColoredPlushieBlock implements EntityBlock {
     @Override