+++ Common/src/main/java/io/github/sirjain0/perfectplushies/client/CommonClientClass.java
@@ -1,5 +1,0 @@
-package io.github.sirjain0.perfectplushies.client;
-
-public class CommonClientClass {
-
-}