+++ Common/src/main/java/io/github/sirjain0/perfectplushies/init/TagInit.java
@@ -1,12 +1,5 @@
 package io.github.sirjain0.perfectplushies.init;
 
-import io.github.sirjain0.perfectplushies.Constants;
-import net.minecraft.core.registries.Registries;
-import net.minecraft.resources.ResourceLocation;
-import net.minecraft.tags.TagKey;
-import net.minecraft.world.item.Item;
-import net.minecraft.world.level.block.Block;
-
 public class TagInit {
 
 //    public static final TagKey<Item> VILLAGE_PLUSHIES_ITEMS = TagKey.create(Registries.ITEM, new ResourceLocation(Constants.MODID, "village_plushies"));