+++ Fabric/src/main/java/io/github/sirjain0/perfectplushies/platform/FabricPlatformHelper.java
@@ -1,11 +1,8 @@
 package io.github.sirjain0.perfectplushies.platform;
 
 import io.github.sirjain0.perfectplushies.platform.services.IPlatformHelper;
-import io.github.sirjain0.perfectplushies.registration.RegistryObject;
 import net.fabricmc.loader.api.FabricLoader;
-import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.Mob;
-import net.minecraft.world.item.SpawnEggItem;
 import software.bernie.geckolib.animatable.GeoItem;
 
 import java.util.function.Consumer;
@@ -35,10 +32,7 @@ public class FabricPlatformHelper<T extends Mob> implements IPlatformHelper<T> {
         return GeoItem.makeRenderer(item);
     }
 
-    @Override
-    public SpawnEggItem createSpawnEggItem(RegistryObject<EntityType<T>> entityTypeRegistryObject, int primaryColor, int secondaryColor) {
-        return new SpawnEggItem(entityTypeRegistryObject.get(), primaryColor, secondaryColor, new SpawnEggItem.Properties());
-    }
+
 
     @Override
     public void registerFabricRenderer(Consumer<Object> consumer) {