+++ Forge/src/main/java/io/github/sirjain0/perfectplushies/platform/ForgePlatformHelper.java
@@ -1,12 +1,7 @@
 package io.github.sirjain0.perfectplushies.platform;
 
 import io.github.sirjain0.perfectplushies.platform.services.IPlatformHelper;
-import io.github.sirjain0.perfectplushies.registration.RegistryObject;
-import net.minecraft.world.entity.EntityType;
-import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.Mob;
-import net.minecraft.world.item.SpawnEggItem;
-import net.minecraftforge.common.ForgeSpawnEggItem;
 import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.loading.FMLLoader;
 import software.bernie.geckolib.animatable.GeoItem;
@@ -42,11 +37,5 @@ public class ForgePlatformHelper<T extends Mob> implements IPlatformHelper<T>{
     public Supplier<Object> getRenderProvider(GeoItem item) {
         return null;
     }
-
-    @Override
-    public SpawnEggItem createSpawnEggItem(RegistryObject<EntityType<T>> entityTypeRegistryObject, int primaryColor, int secondaryColor) {
-        return new ForgeSpawnEggItem(entityTypeRegistryObject, primaryColor,secondaryColor, new SpawnEggItem.Properties());
-    }
-
 
 }